Change stand-alone JS to be in a single file, and better encapsulated.
Now you just add a single thing, the editor_ousupsub object (which is
called Y.M.editor_ousupsub in our code) to the global namespace. This
gives the public API for interacting with our editors.

This change required some knock-on changes to the test scripts, but
everything seems to be working again.