Fix vertical alignments.
This works for editors of the default height (2 rows) and in the OU
theme at least. (In Moodle standard theme it seems to be 1px out.)

Also, it does not work in Windows 10 Edge browser, and sadly Moodle does
not give us a class on <body> for that, so I have ignored that browser
for now.